refactor(events): make shared code generic
